Add test case for modular swi signatures.
Added swim_test.sh: this is a bash script that will run several test
cases and record the output to a log file.
Added a test mini swi in new modular format (3.0) as well as one in
legacy format (1.0).
Added swim_test.ref, the reference output.
At the very end, the test will compare the logfile to the reference.
If there is an error, and the displayed diff context is too small, just vim -d swim_test.log swim_test.ref.
The reference file can also be seen as further examples of usage, or a
quick way to see what was tested and how it all looked like.