Swim (#5)
Update signing tools for working with EOS images that contain multiple rootfs, that is, multiple sub-images, and thus also multiple signatures. 
Such images start with EOS release 4.28.0. 
On hardware with only 4GB of flash, the image will be converted into an image optimized for the small flash (on the fly via the cli command "install source" as it gets downloaded) and its signature will be updated according to the applicable optimization hardware profile (currently: Sand-4GB and Strata-4GB). During "install source" we are sort of decapsulating the image, which results in one of the inner signature becoming the top signature.